requirements.txt
camera==1.3.0
imutils==0.5.4
numpy==1.26.4
Pillow==10.2.0
pyautogui==0.9.54
scikit_learn==1.4.0
tensorflow==2.15.0
tflearn==0.5.0
